警告削除作業途中

+++ fbx/FbxModel.cpp
@@ -16,7 +16,7 @@ void FbxModel::CreateBuffers(ID3D12Device* device)
 	HRESULT result;
 
 	//頂点データ全体のサイズ
-	UINT sizeVB = static_cast<UINT>(sizeof(VertexPosNormalUvSkin)) * vertices.size();
+	UINT sizeVB = static_cast<UINT>(sizeof(VertexPosNormalUvSkin)) * static_cast<UINT>(vertices.size());
 
 	//頂点バッファ生成
 	CD3DX12_HEAP_PROPERTIES v1 = CD3DX12_HEAP_PROPERTIES(D3D12_HEAP_TYPE_UPLOAD);
